arch: arm: Separate common exc header file
This commit Separate exc.h into two header files, 'cortex_a_r/exc.h' and
'cortex_m/exc.h'. Still, keep 'exc.h' so that we don't need to change
other files which include it.
